relocate controlling functions to ledger
