+++ wandering-warriors/modules/__init__.py
@@ -1,6 +1,5 @@
 from .abacus import Abacus
-from .controller import Controller
 from .draw_pad import DrawPad
 from .ledger import Ledger
 
-__all__ = ('Abacus', 'DrawPad', 'Ledger', 'Controller')
+__all__ = ('Abacus', 'DrawPad', 'Ledger')
